Difficulties now reflect my Rock Band builds
Also the ranks can now go as high as 999. I'm not sure what the actual limit is, but I have been able to go up way higher before it stops reading the ranks. Regarding Magma itself, this is the highest number it can render in the rank slider before it bugs out.